refactor: Split up MyAccounts component
